Fix text wrapping across breakpoints.
We were asked to remove the line break between the copy inside the
accordian. The result of this was that the text wrapping didn't
quite look right. After chatting with Harry we decided to try and
wrap the text in line with the buttons. This change to the styling
attempts to match the copy width to the total width of buttons
across all breakpoints.