update thrasher tracking
On request from D&I we update the data-link-name associated with
hiding local news, just so it's clear exactly what any associated
click events are related to. We also update the control to expand
or close the thrasher so we have a unique link name for each of
these actions.